requirements.txt
numpy
scipy
pandas
scikit-learn
matplotlib
torch
torchaudio
torchsummary
jupyter
notebook
ipykernel
tqdm
pyyaml
soundfile
librosa
auraloss
tensorboard
